item dto 및 레포에 long 값 아이디 int로 수정

+++ itemservice/src/main/java/com/ecommerce/itemservice/dto/ItemValues.java
@@ -14,7 +14,7 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ItemValues { // 아이템 간이 정보
-    private long itemId;
+    private int itemId;
     private short category;
     private String itemName;
     private String itemDescription;
